rename Library test project to SharedLibrary (#4985)
