+++ tests/src/Simple/SharedLibrary/SharedLibrary.cmd
