+++ tests/src/Simple/SharedLibrary/no_cpp
